pyproject.toml
[project]
name = "agentic-ai-workflow-and-ai-agent-basic-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.20",
    "langfuse>=3.3.2",
    "loguru>=0.7.3",
    "mistralai>=1.9.9",
    "openai>=1.102.0",
    "openai-agents>=0.2.10",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.11",
]
